+++ src/redturtle/prenotazioni/adapters/booker.py
@@ -16,8 +16,7 @@ from zope.interface import implementer
 
 
 class IBooker(Interface):
-    """ Interface for a booker
-    """
+    """Interface for a booker"""
 
 
 @implementer(IBooker)
@@ -33,11 +32,8 @@ class Booker(object):
     @property
     @memoize
     def prenotazioni(self):
-        """ The prenotazioni context state view
-        """
-        return self.context.unrestrictedTraverse(
-            "@@prenotazioni_context_state"
-        )  # noqa
+        """The prenotazioni context state view"""
+        return self.context.unrestrictedTraverse("@@prenotazioni_context_state")  # noqa
 
     def get_available_gate(self, data_prenotazione, data_scadenza=None):
         """
@@ -55,7 +51,7 @@ class Booker(object):
         return choice(self.prenotazioni.get_less_used_gates(data_prenotazione))
 
     def _create(self, data, duration=-1, force_gate=""):
-        """ Create a Booking object
+        """Create a Booking object
 
         :param duration: used to force a duration. If it is negative it will be
                          calculated using the tipology
@@ -67,9 +63,7 @@ class Booker(object):
         else:
             booking_date = data["booking_date"]
 
-        container = self.prenotazioni.get_container(
-            booking_date, create_missing=True
-        )
+        container = self.prenotazioni.get_container(booking_date, create_missing=True)
         tipology = data.get("tipology", "")
         if duration < 0:
             # if we pass a negative duration it will be recalculated
@@ -92,9 +86,7 @@ class Booker(object):
             "tipologia_prenotazione": data.get("tipology", ""),
         }
         if not force_gate:
-            available_gate = self.get_available_gate(
-                booking_date, data_scadenza
-            )
+            available_gate = self.get_available_gate(booking_date, data_scadenza)
             # if not available_gate: #
             if available_gate is None:
                 # there isn't a free slot in any available gates
@@ -112,9 +104,7 @@ class Booker(object):
             setattr(obj, attribute, at_data[attribute])
 
         # set delete token
-        expiration = datetime.combine(
-            obj.data_prenotazione.date(), time(0, 0, 0)
-        )
+        expiration = datetime.combine(obj.data_prenotazione.date(), time(0, 0, 0))
         token = IDeleteTokenProvider(obj).generate_token(expiration=expiration)
         annotations = IAnnotations(obj)
         annotations[DELETE_TOKEN_KEY] = token.decode("utf-8")
@@ -122,8 +112,12 @@ class Booker(object):
         annotations[VERIFIED_BOOKING] = False
         if not api.user.is_anonymous():
             user = api.user.get_current()
-            fiscalcode = at_data['fiscalcode'].upper()
-            if fiscalcode and (user.getProperty('fiscalcode') or "").upper() == fiscalcode:
+            data_fiscalcode = at_data.get("fiscalcode", "") or ""
+            fiscalcode = data_fiscalcode.upper()
+            if (
+                fiscalcode
+                and (user.getProperty("fiscalcode") or "").upper() == fiscalcode
+            ):
                 logger.info("booking %s verified", at_data)
                 annotations[VERIFIED_BOOKING] = True
 
@@ -142,8 +136,7 @@ class Booker(object):
             return self._create(data, duration=duration, force_gate=force_gate)
 
     def fix_container(self, booking):
-        """ Take a booking and move it to the right date
-        """
+        """Take a booking and move it to the right date"""
         booking_date = booking.getData_prenotazione()
         old_container = booking.aq_parent
         new_container = self.prenotazioni.get_container(